feat: User의 getTotalDrawResult 메소드 구현
 - getDrawResult 메소드 추상화 해제
 - Dealer의 getDrawResult 메소드 변경 -> getFirstDrawResult
