+++ README.md
@@ -24,7 +24,7 @@
     - n: 다음 플레이어 실행
         - [x] 버스트, 블랙잭 여부 확인
             - [x] 현재 점수 확인
-                - [ ] ACE 보유 여부 확인
+                - [x] ACE 보유 여부 확인
 - [x] 딜러 카드 분배
     - [x] 딜러 드로우 여부 확인 (16이하)
         - [x] 버스트, 블랙잭 여부 확인
