+++ src/test/java/domain/user/DealerTest.java
@@ -36,7 +36,7 @@ class DealerTest {
         dealer.draw(new Card(Symbol.CLOVER, Type.EIGHT));
         dealer.draw(new Card(Symbol.DIAMOND, Type.ACE));
 
-        assertThat(dealer.getDrawResult()).isEqualTo("딜러카드: 8클로버");
+        assertThat(dealer.getFirstDrawResult()).isEqualTo("딜러카드: 8클로버");
     }
 
     @ParameterizedTest
